% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotmap.R
\name{break_breaks}
\alias{break_breaks}
\title{Tick (break) positions}
\usage{
break_breaks(levels, by = 50)
}
\arguments{
\item{levels}{chr vector of (x or y) variables in correct order}

\item{by}{tick spacer length}
}
\value{
chr vector of break/tick positions to be displayed
}
\description{
Tick (break) positions
}
