% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotmap.R
\name{hotmap}
\alias{hotmap}
\title{Plot hotmap}
\usage{
hotmap(mat, melted = FALSE, return.data = FALSE, x = NULL, y = NULL,
  fill = NULL, x.order = TRUE, y.order = TRUE, order = TRUE,
  limits = c(-10, 10), cols = colours(), title = "",
  x.breaks = waiver(), x.labels = waiver(), x.name = NULL,
  x.labels.by = NULL, x.labels.numeric = FALSE, y.breaks = waiver(),
  y.labels = waiver(), y.name = NULL, y.labels.by = NULL,
  y.labels.numeric = FALSE, title.hjust = 0.5, legend.position = "none",
  panel.border = element_blank(), margins = c(0, 0, 0, 0), ratio = 1,
  x.ticksize = 6, y.ticksize = 6, bl.label = "", br.label = "",
  tl.label = "", tr.label = "", label.size = 0.1)
}
\arguments{
\item{mat}{matrix to be plotted}

\item{melted}{TRUE if matrix in tidy format}

\item{x}{axis}

\item{y}{axis}

\item{fill}{value}

\item{x.order}{order of variables on x axis}

\item{y.order}{order of variables on x axis}

\item{order}{do not x.order or y.order if FALSE}

\item{limits}{colour limits}

\item{cols}{colours}

\item{title}{plot}

\item{x.breaks}{tick mark positions}

\item{x.labels}{tick mark labels}

\item{x.name}{axis label}

\item{x.labels.by}{tick spaces}

\item{x.labels.numeric}{convert labels to numbering}

\item{y.breaks}{tick mark positions}

\item{y.labels}{tick mark labels}

\item{y.name}{axis label}

\item{y.labels.by}{tick spaces}

\item{y.labels.numeric}{convert labels to numbering}

\item{title.hjust}{adjust title position. 0.5 is middle}

\item{legend.position}{hide or display key and if so where}

\item{panel.border}{show plot border, if so which}

\item{margins}{border space around plot: c(x,x,x,x)}

\item{ratio}{aspect ratio of x and y axis}

\item{x.ticksize}{tick label size}

\item{y.ticksize}{tick label size}

\item{bl.label}{add text on plot bottom left}

\item{br.label}{add text on plot bottom right}

\item{tl.label}{add text on plot top left}

\item{tr.label}{add text on plot top right}

\item{label.size}{text size}
}
\description{
Plot hotmap
}
