% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotmap.R
\name{hotmap_break_label}
\alias{hotmap_break_label}
\title{Breaks and labels}
\usage{
hotmap_break_label(levels, breaks, labels, by = 50, labels.numeric = FALSE)
}
\arguments{
\item{levels}{vector of (x or y) variables in correct axis order}

\item{breaks}{when break positions unchanged, same as positions in <levels>}

\item{labels}{when labels unchanged, same as names in <levels>}

\item{by}{tick spacer length}

\item{labels.numeric}{replace labels with numbering}
}
\value{
list of two character vectors: breaks and labels
}
\description{
Breaks and labels
}
