% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotmap.R
\name{hotmap_order}
\alias{hotmap_order}
\title{Ordering matrix}
\usage{
hotmap_order(mat, x.order = TRUE, y.order = TRUE, order = TRUE)
}
\arguments{
\item{mat}{matrix to be plotted}

\item{x.order}{if chr vect: order with this. if TRUE: generate order. if F: do nothing.}

\item{y.order}{if chr vect: order with this. if TRUE: generate order. if F: do nothing.}

\item{order}{if FALSE, do not generate orders.}
}
\value{
(ordered) matrix
}
\description{
Ordering matrix
}
