% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotmap.R
\name{hotmap_prep}
\alias{hotmap_prep}
\title{Data preparation for plotting}
\usage{
hotmap_prep(mat, x.order = TRUE, y.order = TRUE, order = TRUE)
}
\arguments{
\item{mat}{matrix to be plotted}

\item{x.order}{if chr vect: order with this. if TRUE: generate order. if F: do nothing.}

\item{y.order}{if chr vect: order with this. if TRUE: generate order. if F: do nothing.}

\item{order}{if FALSE, do not generate orders.}
}
\value{
(ordered) melted matrix
}
\description{
Data preparation for plotting
}
