% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotmap.R
\name{label_breaks}
\alias{label_breaks}
\title{Tick labels}
\usage{
label_breaks(until, by = 50)
}
\arguments{
\item{until}{length of levels vector}

\item{by}{tick spacer length}
}
\value{
chr vector of tick labels to be displayed
}
\description{
Tick labels
}
